feat(pooler): refactor package name for clarity
