+++ go/multipooler/poolerserver/plugin_pq.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package pooler
+package poolerserver
 
 import (
 	_ "github.com/lib/pq" // PostgreSQL driver
